# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/me": {
      "get": {
        "operationId": "getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns public user profile data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the currently authenticated user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns an access token and a refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairDto"
                }
              }
            }
          },
          "401": {
            "description": "Wrong username or password"
          }
        },
        "summary": "Log in and receive an access + refresh token pair",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "description": "Requires a valid `REGISTER_SECRET` env value in the request body to prevent open registration. Returns the generated activation hash (to be sent via email in a real deployment).",
        "operationId": "register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created — returns the activation hash in dev mode"
          },
          "401": {
            "description": "Invalid register secret"
          },
          "409": {
            "description": "Username or email already taken"
          }
        },
        "summary": "Register a new user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/activate": {
      "get": {
        "description": "Verifies the activation hash and marks the account as active. Returns an access + refresh token pair so the user is immediately logged in after activation.",
        "operationId": "activateAccount",
        "parameters": [
          {
            "name": "hash",
            "required": true,
            "in": "query",
            "description": "The MD5 activation hash from the registration response",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account activated — returns an access + refresh token pair",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairDto"
                }
              }
            }
          },
          "404": {
            "description": "Invalid or already-used activation hash"
          }
        },
        "summary": "Activate a user account",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "Refresh tokens are single-use — every call rotates in a new one, invalidating the old one.",
        "operationId": "refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns a new access + refresh token pair",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid, expired, or already-used refresh token"
          }
        },
        "summary": "Exchange a refresh token for a new access + refresh token pair",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/api-keys": {
      "get": {
        "operationId": "listApiKeys",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApiKeyDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the current user's API keys",
        "tags": [
          "api-keys"
        ]
      },
      "post": {
        "description": "The raw key is only ever returned in this response — store it now, it cannot be retrieved again.",
        "operationId": "createApiKey",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedApiKeyDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Generate a new API key",
        "tags": [
          "api-keys"
        ]
      }
    },
    "/auth/api-keys/{id}": {
      "delete": {
        "operationId": "revokeApiKey",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke an API key",
        "tags": [
          "api-keys"
        ]
      }
    },
...
  },
  "info": {
    "title": "MCP Bridge API",
    "description": "Relay backend exposing desktop-agent MCP servers over HTTP",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "MeDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "subscription": {
            "type": "string"
          },
          "isActivated": {
            "type": "boolean"
          }
        },
        "required": [
          "username",
          "role",
          "subscription",
          "isActivated"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "example": "admin"
          },
          "password": {
            "type": "string",
            "example": "hunter2"
          }
        },
        "required": [
          "user",
          "password"
        ]
      },
      "TokenPairDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Short-lived JWT used to authenticate API requests"
          },
          "refreshToken": {
            "type": "string",
            "description": "Long-lived opaque token used to obtain a new access token via /auth/refresh"
          }
        },
        "required": [
          "accessToken",
          "refreshToken"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "admin",
            "description": "Unique username (lowercase alphanumeric / underscore, no hyphens)"
          },
          "password": {
            "type": "string",
            "example": "hunter2",
            "description": "Password (min 8 chars)"
          },
          "registerSecret": {
            "type": "string",
            "example": "SUPER_SECRET_REGISTER_KEY",
            "description": "Server-side registration secret from env REGISTER_SECRET"
          }
        },
        "required": [
          "username",
          "password",
          "registerSecret"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "admin"
          },
          "refreshToken": {
            "type": "string",
            "description": "The refresh token issued at login/activation"
          }
        },
        "required": [
          "username",
          "refreshToken"
        ]
      },
      "ApiKeyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "prefix": {
            "type": "string",
            "description": "First few characters of the key, for identification — the rest is never shown again"
          },
          "createdAt": {
            "type": "string"
          },
          "lastUsedAt": {
            "type": "string",
            "nullable": true
          },
          "revokedAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "prefix",
          "createdAt",
          "lastUsedAt",
          "revokedAt"
        ]
      },
      "CreateApiKeyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "CI server",
            "description": "A label to tell this key apart from others in the list"
          }
        },
        "required": [
          "name"
        ]
      },
      "CreatedApiKeyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "prefix": {
            "type": "string",
            "description": "First few characters of the key, for identification — the rest is never shown again"
          },
          "createdAt": {
            "type": "string"
          },
          "lastUsedAt": {
            "type": "string",
            "nullable": true
          },
          "revokedAt": {
            "type": "string",
            "nullable": true
          },
          "key": {
            "type": "string",
            "description": "The full raw API key — shown once, at creation time, and never retrievable again"
          }
        },
        "required": [
          "id",
          "name",
          "prefix",
          "createdAt",
          "lastUsedAt",
          "revokedAt",
          "key"
        ]
      },
...
    }
  }
}